Remove a comment suggesting to mocking the state
At some point we could expose the log for mocking
the test arrangement. As is, the stateful tests are
more complicated than they need to be.

Since the log package shares the server package and
we aren't importing a dependency we can ignore it
for now. Someone could argue that the verbosity
helps convey the intention.

We will see where the curriculum goes, this might
need to change.